feat(clock): implement clock generation and BPM control in clockEngine.hpp moving from main
